+++ src/test/java/roomescape/dao/FakeReservationDaoImpl.java
@@ -1,10 +1,9 @@
-package roomescape.domain.dao;
+package roomescape.dao;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
-import roomescape.dao.ReservationDao;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationDate;
 import roomescape.exception.InvalidReservationException;
