+++ src/test/java/roomescape/dao/FakeReservationTimeDaoImpl.java
@@ -1,11 +1,10 @@
-package roomescape.domain.dao;
+package roomescape.dao;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
-import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
 
 public class FakeReservationTimeDaoImpl implements ReservationTimeDao {
